requirements.txt
confluent_kafka
pymongo
python-dotenv
configparser